% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_label_integration.R
\name{celltype_label_integration}
\alias{celltype_label_integration}
\title{Integrate unsupervised and supervised (immune-celltype) cluster labels}
\usage{
celltype_label_integration(
  metadata,
  adjacency_mat,
  cellid_colname = "cell_ID",
  unsupervised_colname = "clust",
  supervised_colname = "celltype_granular",
  supervised_labels_keep = unique(c(names(HieraType::markerslist_immune),
    names(HieraType::markerslist_tcellmajor), names(HieraType::markerslist_cd4tminor),
    names(HieraType::markerslist_cd8tminor))),
  dissolve_smallcluster_if_overwritten_prop_greaterthan = c(0.9, 0.5),
  and_dissolve_smallcluster_if_finalcluster_prop_lessthan = c(0.05, 0.01)
)
}
\arguments{
\item{metadata}{data.table or data.frame of metadata containing at least c("cell_ID", unsupervised_colname, supervised_colname) columns}

\item{adjacency_mat}{cells x cells sparse similarity matrix used to determine how to dissolve rare clusters
(When a unsupervised cluster is predominantly immune, but HieraType doesnt' label it as such, then there may be leftover cells -
these get relabeled to the most common cluster call amongst that cell's neighbors).  Recommended to use the same adjacency matrix used to generate HieraType results.}

\item{unsupervised_colname}{colname for the unsupervised (i.e., leiden) clusters}

\item{supervised_colname}{colname for the supervised (hieratype) clusters}

\item{supervised_labels_keep}{HieraType cell type categories (i.e., immune celltypes) that should be kept / override the unsupervised labels if they are present.}

\item{dissolve_smallcluster_if_overwritten_prop_greaterthan}{This argument is paired with \code{and_dissolve_smallcluster_if_finalcluster_prop_lessthan}.
For an unsupervised cluster which is mostly overwritten (i.e., at least 90\% of cells are re-labeled to HieraType), then re-assign the cells in that
cluster to the most common label amongst similar neighboring cells in the adjacency matrix IF the final cluster proportion is sufficiently small.
(i.e., less  than 5\% of all cells.).
The defaults below set up two conditions for re-assigning cells in rare unsupervised clusters which were mostly designated to a HieraType cluster.
Example:
dissolve_smallcluster_if_overwritten_prop_greaterthan = c(0.9, 0.5)
and_dissolve_smallcluster_if_finalcluster_prop_lessthan = c(0.05, 0.01)
This means that cells in for particular unsupervised clusters would be re-assigned if
$>$90\% were overwritten by hieratype labels and the final unsupervised cluster proportion < 0.05 or if
$>$50\% were overwritten by hieratype labels and the final unsupervised cluster proportion < 0.01.
As shown in the example, these two arguments should be numeric vectors of equal length.}

\item{and_dissolve_smallcluster_if_finalcluster_prop_lessthan}{This argument is paired with \code{dissolve_smallcluster_if_overwritten_prop_greaterthan}.
See notes above for guidance.}
}
\value{
a data.table with ('cellid_colname', 'unsupervised_colname', 'supervised_colname', 'celltype', and 'dissolved_cell'),
where celltype is the new integrated label and dissolved_cell indicates whether the cell came from a dissolved cluster and was re-assigned using nearest neighbors in the adjacency matrix.
}
\description{
Integrate unsupervised and supervised (immune-celltype) cluster labels
}
