% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_postprob_tables.R
\name{combine_postprob_tables}
\alias{combine_postprob_tables}
\title{Utility function for making / remaking or restyling
combined posterior probability tables from clustering models (possibly at different thresholds).}
\usage{
combine_postprob_tables(
  pipeline,
  models,
  celltype_call_threshold = 0.5,
  return_all_columns = FALSE
)
}
\arguments{
\item{pipeline}{a pipeline object created using \code{make_pipeline()}}

\item{models}{list of models, created using \code{run_pipeline}}

\item{celltype_call_threshold}{Used to determing the celltype_thresh column in returned table.
This column shows the most granular cell type with posterior probability > \code{celltype_call_threshold} for each cell.}

\item{return_all_columns}{if FALSE (default), columns with posterior probability scores for
individual cell type classes are dropped from the returned table.}
}
\description{
Utility function for making / remaking or restyling
combined posterior probability tables from clustering models (possibly at different thresholds).
}
