% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_mvn_mixture.R
\name{em_mvn_mixture}
\alias{em_mvn_mixture}
\title{Fit multivariate gaussian mixture to celltype metagene scores}
\usage{
em_mvn_mixture(
  scores,
  initmu,
  initsigma,
  maxiter = 1000,
  rel.eps.convergence = 1e-06,
  component.variance = TRUE,
  prior_prob_level = NULL,
  verbose = 2,
  min_pos_score_mean = 0.1,
  max_neg_score_mean = 0,
  intercepts = NULL
)
}
\arguments{
\item{scores}{a cells x components matrix of metagene scores}

\item{initmu}{optional components x components matrix of initial mean values}

\item{initsigma}{optional components length list of components x components covariance matrices used for starting values.}

\item{maxiter}{maximum iterations before quitting em algorithm.}

\item{rel.eps.convergence}{if (abs(ll\link{iter} - ll\link{iter-1}) / ll\link{iter}) < rel.eps.convergence, then algorithm is converged.}

\item{component.variance}{FALSE is an experimental option.  This should typically be the default TRUE.}

\item{prior_prob_level}{optional vector of posterior probability weights.}

\item{verbose}{integer in range of 0-4 controlling verbosity of printed messages.}

\item{min_pos_score_mean}{minimum value for}

\item{max_neg_score_mean}{maximum value for the mean of off-component metagenes.}
}
\description{
Fit multivariate gaussian mixture to celltype metagene scores
}
