% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_frequency.R
\name{gene_frequency}
\alias{gene_frequency}
\title{Compute the frequency of each gene from a sparse counts matrix}
\usage{
gene_frequency(
  x,
  obs = NULL,
  batch_variable = NULL,
  cellid_colname = "cell_ID"
)
}
\arguments{
\item{x}{A sparse genes x cells counts matrix}

\item{obs}{A data frame of metadata information, containing columns for the 'batch variable' and 'cell id'.
Or, row names of the data frame indicating 'cell id'.}

\item{batch_variable}{column in \code{obs} indicating 'batch'.  i.e., 'patient' or 'samplename' or 'slide_id'.}

\item{cellid_colname}{column in \code{obs} indicating the cell identifier, i.e., 'cell_ID'.}
}
\description{
This function computes the frequency of each gene from the provided sparse counts matrix \code{x}.
Returned is a gene-named vector of frequencies which sum to 1.
If obs, batch_variable, and cellid_colname are provided, these frequencies are calculated separately per batch,
returning a genes x batches matrix with columns summing to 1.
}
