% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_sketch_get_plaid.R
\name{geo_sketch_get_plaid}
\alias{geo_sketch_get_plaid}
\title{Bin cells into "plaids"}
\usage{
geo_sketch_get_plaid(
  X,
  N = sqrt(nrow(X)),
  alpha = 0.05,
  max_iter = 200,
  return_grid = FALSE
)
}
\arguments{
\item{X}{feature matrix with cell IDs as rows and feature IDs as columns ()}

\item{N}{desired sample size}

\item{alpha}{divide feature space into uniform 'plaid' bins, where number of bins is in the range of
(\verb{N * (1-alpha),  N * (1+alpha)}), where}

\item{max_iter}{maximum number of iterations used to achieve an acceptable
minimum number of bins}

\item{return_grid}{determines whether or not to pass back bin labels instead of just the sample indices.}
}
\description{
R implementation of algorithm found in 'geosketch' python package
}
