% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene_coefficients_plot.R
\name{metagene_coefficients_plot}
\alias{metagene_coefficients_plot}
\title{Plot the regression coefficients of metagene scores}
\usage{
metagene_coefficients_plot(
  metagenes,
  label_cutoff_abscoef = 0.05,
  label_cutoff_n = 100,
  max.overlaps = 100,
  annotate_nonzero = TRUE
)
}
\arguments{
\item{metagenes}{metagenes object created by fit_metagene_scores}

\item{label_cutoff_abscoef}{minimum absolute value of regression coefficient to be labeled with geom_label_repel}

\item{label_cutoff_n}{maximum number of predictors to annotate with geom_label_repel}

\item{max.overlaps}{max.overlaps argument passed to ggrepel::geom_label_repel for labeling predictor names .}

\item{annotate_nonzero}{if TRUE, create a text annotation showing the fraction of non-zero predictors out of all possible.}
}
\description{
Plot the regression coefficients of metagene scores
}
