% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene_pairsplot.R
\name{metagene_pairsplot}
\alias{metagene_pairsplot}
\title{pairsplot for metagene scores and classifications using GGally and ggplot2}
\usage{
metagene_pairsplot(
  metagenes,
  var,
  nsamples = NULL,
  wts = NULL,
  colorvar = NULL,
  wtlim = 0.5,
  ptalpha = 0.5,
  ptsize = 0.1,
  obs = NULL
)
}
\arguments{
\item{metagenes}{a metagenes object created using \code{fit_metagene_scores()}}

\item{var}{one of 'ypost', 'yhat', or 'y' to be plotted (reflecting posterior-averaged, model-predicted, or raw scores of the index gene, respectively).}

\item{nsamples}{if NULL, plots all cells.  Otherwise, plots a random sample of \code{nsamples} cells (quicker for large datasets).}

\item{wts}{optional cells-length vector of wts, reflecting prior probability that cell belongs to \strong{any} of the possible metagene clusters.}

\item{colorvar}{Optional categorical variable by which to color the cells.  This should be provided as a .}

\item{wtlim}{optional lower weight limit for plotting cells.  If \code{wts} are provided, limit the plot to cells where \code{wts > wtlim}.}

\item{ptalpha}{alpha value controlling transparency of plotted points.}

\item{ptsize}{controls size of plotted points.}

\item{obs}{optional cell-level data frame / data.table that is inner-joined onto the metagene scores.
This can be used to subset the cells plotted (by only passing \code{obs} for the cells you want to plot) and/or including \code{colorvar} variable to be used for plotting.}
}
\description{
Make a pairsplot for metagene scores and classifications using GGally and ggplot2.
}
