% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_lincombs_contrasts.R
\name{prepare_lincombs}
\alias{prepare_lincombs}
\title{Function used to summarize models with emmeans based contrasts and a model summary.}
\usage{
prepare_lincombs(dat = NULL, fitmethod, original_formula, fam, inlanames)
}
\arguments{
\item{mod}{a model.}

\item{modtime}{optional. An object produced by 'system.time', if provided it is added to the model_summary in returned object.}

\item{groupVar}{optional groupVar.
If specified, this forces pairwise contrasts to be conducted by this variable, even if there are 20 or more levels.}
}
\value{
a list of lists containing data.tables of emmeans summaries.  Intended to be primarily accessed through the smiDE::results function.
}
\description{
Function used to summarize models with emmeans based contrasts and a model summary.
}
