% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_init.R
\name{quantile_init}
\alias{quantile_init}
\title{function to get starting values for cluster means and covariances}
\usage{
quantile_init(scores, q = 0.9, minqu = 0.05)
}
\arguments{
\item{scores}{cells x metagenes matrix of scores}

\item{q}{upper quantile used to get starting means/covariances for initializing clusters}

\item{minqu}{set the score to this value if the 'q'th quantile is below it.}
}
\description{
function to get starting values for cluster means and covariances
}
