% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rowSDs}
\alias{rowSDs}
\title{Calculate row standard deviations without converting to dense matrix.}
\usage{
rowSDs(x)
}
\arguments{
\item{x}{A sparse dgCMatrix}
}
\description{
Calculate row standard deviations without converting to dense matrix.
}
\examples{
 
set.seed(30)
sm <- Matrix::sparseMatrix(i=sample(1:30,450,replace=TRUE),
                           j=sample(1:20,450,replace=TRUE),x=runif(450))
all.equal(rowSDs(sm), apply(sm, 1, sd))

}
