% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Run a hierarchical cell typing pipeline}
\usage{
run_pipeline(
  pipeline,
  counts_matrix,
  adjacency_matrix = NULL,
  initial_prior_weights = NULL,
  celltype_call_threshold = 0.5,
  return_all_columns_postprobs = FALSE,
  ...
)
}
\arguments{
\item{pipeline}{object created using \code{make_pipeline()}}

\item{counts_matrix}{a cells x genes expression matrix.
If modeling scaled pearson residuals (pearson.response = TRUE  by default)
,this should be a raw counts matrix.}

\item{adjacency_matrix}{an optional cells x cells matrix matrix of weights denoting similarity between pairs of cells.
For example, one could use the graph of smoothed nearest neighbors distances used to optimize UMAP embeddings.}

\item{initial_prior_weights}{an optional vector of weights (typically in range of 0-1),
denoting confidence that the cells belong to any of the classes in provided markerslist.
This argument can be used to specify weights for the \strong{first} celltyping step.}

\item{celltype_call_threshold}{Posterior probability threshold used to determine level of calling for \code{celltype_thresh} annotation in the returned \code{post_probs} data.table.
See \code{combine_postprob_tables()} function; these tables can be remade quickly at different thresholds without the need to rerun a pipeline.}

\item{return_all_columns_postprobs}{Whether to return individual posterior probabilities for every celltype in the \code{post_probs} data.table.
See \code{combine_postprob_tables()} function; these columns can be returned later if needed without needing to rerun a pipeline.}

\item{...}{Other arguments to be passed to \code{fit_metagene_scores} and/or \code{cluster_metagenes} functions.}
}
\value{
A list containing:
\itemize{
\item \code{post_probs} - Combined posterior probability tables from all pipeline stages
\item \code{models} - List of fitted clustering models for each pipeline stage
\item \code{metagene_scores} - List of metagene score fits for each pipeline stage
}
}
\description{
Run a hierarchical cell typing pipeline
}
