% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_quasipoisson_pca_seurat-copy.R,
%   R/sparse_quasipoisson_pca_seurat.R
\name{sparse_quasipoisson_pca_seurat}
\alias{sparse_quasipoisson_pca_seurat}
\title{Run sparse PCA Using Quasi-poisson pearson residuals (skip dense normalization)}
\usage{
sparse_quasipoisson_pca_seurat(
  x,
  totalcounts = NULL,
  grate = NULL,
  scale.max = 10,
  reduction.key = "PC_",
  do.scale = TRUE,
  do.center = TRUE,
  quasi_poisson_variance_inflation = 1.01,
  npcs = 50,
  ncores = 1,
  verbose = TRUE,
  block.size = 10^4,
  ndims.print = 1:5,
  nfeatures.print = 30,
  return_seurat_reduction = TRUE,
  estimate_dispersion_from_poisson = FALSE,
  only_return_sds = FALSE
)

sparse_quasipoisson_pca_seurat(
  x,
  totalcounts = NULL,
  grate = NULL,
  scale.max = 10,
  reduction.key = "PC_",
  do.scale = TRUE,
  do.center = TRUE,
  quasi_poisson_variance_inflation = 1.01,
  npcs = 50,
  ncores = 1,
  verbose = TRUE,
  block.size = 10^4,
  ndims.print = 1:5,
  nfeatures.print = 30,
  return_seurat_reduction = TRUE,
  estimate_dispersion_from_poisson = FALSE,
  only_return_sds = FALSE
)
}
\arguments{
\item{x}{A sparse genes x cells counts matrix}

\item{totalcounts}{The total UMI counts across all gene targets in each cell.  This can be pre-computed by the user and passed in cases where the counts matrix \code{x} contains only a
subset (of, say, highly variable) genes to be used in PCA.}

\item{grate}{The gene frequencies.  This can be pre-computed by the user and passed in cases where the counts matrix \code{x} contains only a
subset (of, say, highly variable) genes to be used in PCA.}

\item{scale.max}{clip expression more than 'scale.max' standard deviations from the mean.
Analogous to Seurat::ScaleData scale.max argument, with same default of "10".}

\item{reduction.key}{Seurat key}

\item{do.scale}{gene x gene cov matrix reflects as if pearson residuals are scaled by their standard deviation}

\item{do.center}{gene x gene cov matrix reflects as if the pearson residuals were centered by their mean}

\item{quasi_poisson_variance_inflation}{factor of overdispersion for Variance relative to the mean.  i.e., V(x) = E(x) * phi, where phi is the variance inflation factor}

\item{npcs}{the # of PCs}

\item{ncores}{the # of cores to use to project embeddings}

\item{verbose}{print out seurat-style PCA summary}

\item{block.size}{the # of cells to project into PCA space at a time. Can reduce block.size for lower-memory computing.}

\item{ndims.print}{the # of first PC's to print for verbose-seurat-style printout.}

\item{nfeatures.print}{the # of top genes to show per PC for verbose-seurat-style printout.}

\item{return_seurat_reduction}{if TRUE, returns a reduction.data object in Seurat format.}
}
\description{
Recreate Seurat RunPCA output from pearson residual normalization (described in Lause 2021),
WITHOUT creating a dense cells x genes scaled data matrix.
This function uses Quasi-poisson instead of Negative Binomial distribution for computational convenience.
Computes PCA via SVD decomposition of genes x genes matrix instead of
SVD decomposition of cells x genes dense matrix, which is useful for memory when cells x genes >> genes^2 .

Recreate Seurat RunPCA output from pearson residual normalization (described in Lause 2021),
WITHOUT creating a dense cells x genes scaled data matrix.
This function uses Quasi-poisson instead of Negative Binomial distribution for computational convenience.
Computes PCA via SVD decomposition of genes x genes matrix instead of
SVD decomposition of cells x genes dense matrix, which is useful for memory when cells x genes >> genes^2 .
}
