% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy_kmeans_clusters.R
\name{xy_kmeans_clusters}
\alias{xy_kmeans_clusters}
\title{K-means cluster in x-y space, used for downstream spatial random effect models.}
\usage{
xy_kmeans_clusters(
  metadata,
  x_coord_col = "sdimx",
  y_coord_col = "sdimy",
  cluster_suffix = "_cluster",
  cluster_name = "k_cluster",
  k = NULL,
  k_prop_n = 0.25,
  seed = 1,
  split_neighbors_by_colname = NULL
)
}
\arguments{
\item{metadata}{metadata which includes cell-level x and y coordinates
with names corresponding to \code{x_coord_col} and \code{y_coord_col} arguments.}

\item{x_coord_col}{name of column with x-coordinate for cell, i.e., "sdimx".}

\item{y_coord_col}{name of column with x-coordinate for cell, i.e., "sdimy".}

\item{cluster_suffix}{output x,y cluster centroids with this suffix,
i.e., "_cluster" => "sdimx_cluster", "sdimy_cluster"}

\item{cluster_name}{name for column to indicate each unique k-means cluster.}

\item{k}{number of kmeans clusters}

\item{k_prop_n}{number of kmeans clusters as a function of total number of cells.
i.e., with k_prop_n = 0.2 and 100 total cells, return 20 clusters.}

\item{seed}{seed set for ensuring reproducible k-means results.
.Random.seed state is restored within the function, so global environment is not affected.}

\item{split_neighbors_by_colname}{In the case of multiple flow-cells, with overlapping range of x and y-coordinates,
specifying this column ensures that different samples/flow-cells with similar x/y coordinates will not belong to the same cluster.}
}
\description{
K-means cluster in x-y space, used for downstream spatial random effect models.
}
