% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_metagenes.R
\name{cluster_metagenes}
\alias{cluster_metagenes}
\title{Cluster metagene scores using a mixture of overfitted gaussian mixture models.}
\usage{
cluster_metagenes(
  metagenes = NULL,
  scores = NULL,
  to_model = c("yhat", "ypost"),
  discourage_double_positive = list(plasma = "immune", endothelial = "immune", epithelial
    = "immune", fibroblast = "immune", smooth_muscle = "immune", mast = "tcell"),
  prior_prob_level = NULL,
  fit_single_positive_only = FALSE,
  fit_threshold_prior_prob_level = 0.1,
  verbose = 1,
  k_components = 6,
  training_cellids = NULL,
  seed = 123
)
}
\arguments{
\item{metagenes}{metagenes object created by}

\item{scores}{optional cells x cell types matrix of metagene scores, only needed if \code{metagenes} not provided.}

\item{to_model}{either 'yhat' (predicted scores, recommended) or 'ypost' posterior-mean metagene scores.}

\item{discourage_double_positive}{list in the format of
list("mycelltype" = c("celltypeA_shouldnt_be_positive_in_mycelltype", "celltypeB_shouldnt_be_positive_in_mycelltype"))
In this example, cells with positive metagene scores for "celltypeA" and/or "celltypeB" will not be used as anchor cells in "mycelltype"
This effectively encourages double positive cells to be assigned to "celltypeA" or "celltypeB" in the event they are double-positive with "mycelltype".
This can be useful if marker genes for "celltypeA","celltypeB" are weak or low expressors, and/or mycelltype has a high expressing promiscuous marker gene which may bleed into other celltypes.}

\item{prior_prob_level}{optional vector of weights / posterior probabilities (likely from a previous cluster_metagenes model)
that the cells belong to \emph{any} of the classes in 'metagenes'.}

\item{fit_single_positive_only}{If TRUE, then models will be trained using cells which are \emph{only} positive for that corresponding cell type.
Otherwise, cells which are positive in both observed AND fitted values for a particular celltype metagene will be included in model training
(even if they have positive scores for other cell types), provided that they pass any \code{discourage_double_positive} constraints.}

\item{fit_threshold_prior_prob_level}{If prior_prob_level is provided, this is the minimum threshold considered for a cell to be included in
mixture model training for a particular cell type.}

\item{k_components}{Number of mixture components to be used in fitting mixture model for each cell type.
Goal is to have 'enough', to flexibly model the distribution of metagene scores.}

\item{training_cellids}{If provided, the metagenes or scores object will be subset to only this set of cells before clustering and scoring.}

\item{seed}{seed to be set in local environment before using k-means to initialize starting parameters.  Ensures consistency after re-running.
Global seed is reset on function exit.}
}
\value{
A list containing:
\itemize{
\item \code{models} - fitted mixture models for each cell type
\item \code{post_probs} - data.table of posterior probabilities with cell_ID, best_class, and best_score
\item \code{llmat} - matrix of log-likelihoods
}
}
\description{
Cluster metagene scores using a mixture of overfitted gaussian mixture models.
}
