% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterwise_foldchange_metrics.R
\name{clusterwise_foldchange_metrics}
\alias{clusterwise_foldchange_metrics}
\title{Compute fold-change, proportion of cells >0, average expression metrics for each gene in each cluster
(comparing with all other clusters)}
\usage{
clusterwise_foldchange_metrics(
  counts = NULL,
  normed = NULL,
  totalcounts = NULL,
  metadata,
  cluster_column,
  cellid_column = "cell_ID"
)
}
\arguments{
\item{counts}{genes x cells matrix of counts.  this only needs to be provided if \code{normed} is not specified.
When \code{normed} is not specified, \code{counts} will be 'totalcount-normalized' before
foldchange metrics are calculated.}

\item{normed}{genes x cells matrix of data to be used for calculating fold change statistics by clusters.
If this is not provided, counts will be 'totalcount-normalized'.
If you desire to use raw counts for computing fold change, use like \code{normed=my_raw_counts}.}

\item{totalcounts}{optional user-specified vector of totalcounts used for normalizing the counts matrix.
(useful if a subsetted counts matrix is passed)}

\item{metadata}{metadata including (at minimum) 'cellid_column' and a 'cluster_column'}

\item{cluster_column}{column in metadata corresponding to cell type}

\item{cellid_column}{column in metadata corresponding to cell id. Should also correspond to the column names of the \code{counts} and \code{normed} expression matrices.}
}
\value{
A data.table with columns: cluster, cluster_expr, clusterprime_expr, gene, cluster_prop,
clusterprime_prop, ncells, typ, fold_change, and fold_change_prop.
}
\description{
This function computes metrics by gene by cluster which can be passed as input to the \code{marker_heatmap()} function.
}
