% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_frequency.R
\name{gene_frequency}
\alias{gene_frequency}
\title{Compute gene frequency from counts matrix.}
\usage{
gene_frequency(
  x,
  obs = NULL,
  batch_variable = NULL,
  cellid_colname = "cell_ID"
)
}
\arguments{
\item{x}{A sparse genes x cells counts matrix}

\item{obs}{A data frame of metadata information, containing columns for the 'batch variable' and 'cell id'.
Or, row names of the data frame indicating 'cell id'.}

\item{batch_variable}{column in \code{obs} indicating 'batch'.  i.e., 'patient' or 'samplename' or 'slide_id'.}

\item{cellid_colname}{column in \code{obs} indicating the cell identifier, i.e., 'cell_ID'.}
}
\description{
Compute gene frequency (possibly by batch) , used in calculation of pearson residuals.
}
