% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_heatmap.R
\name{marker_heatmap}
\alias{marker_heatmap}
\title{Make a marker gene heatmap showing fold changes and proportion of cells expressing the marker gene by cluster.}
\usage{
marker_heatmap(
  foldchange_metrics,
  extras = c(),
  featsuse = NULL,
  topn = 5,
  geneorder = NULL,
  clusterorder = NULL,
  fold_change_min = -1,
  cluster_prop_min = 0.05,
  colordervar = "scaled_fold_change",
  orient_diagonal = TRUE,
  topleft_to_bottomright = FALSE
)
}
\arguments{
\item{foldchange_metrics}{metrics calculated using \code{clusterwise_foldchange_metrics()} function.}

\item{extras}{if provided, these extra genes will be included in the plot, even if they are not in the \code{topn} in terms of
\code{foldchange} for a particular cluster based on fold change or don't meet the \code{cluster_prop_min} requirement.}

\item{featsuse}{if provided, these are the (only) genes that will be used , in addition to any \code{extras}}

\item{topn}{plot the \code{topn} genes (in terms of fold_change, which are expressed in at least cluster_prop_min pct of cells) from each cluster.}

\item{geneorder}{if provided, this allows to manually control the order of genes.
Probably should only used if the exact set of features is also provided with \code{featsuse}.}

\item{clusterorder}{if provided this allows to manually control the order in which clusters are plotted.}

\item{fold_change_min}{minimum fold change, order to qualify to be plotted.}

\item{cluster_prop_min}{minimum proportion of cells that the gene must be expressed in, in order to qualify to be plotted.}

\item{orient_diagonal}{Orient marker genes from bottom-left to top-right (usually easier to read).}

\item{topleft_to_bottomright}{if TRUE, and orient_diagonal also TRUE, do TL to BR instead of BL to TR.}

\item{colorordervar}{column in foldchange_metrics used in ComplexHeatmap::Heatmap function, which
determines row and column positions for clusters and genes.}
}
\value{
A ggplot2 object showing the marker heatmap.
}
\description{
Make a marker gene heatmap showing fold changes and proportion of cells expressing the marker gene by cluster.
}
